requirements-ops.txt
fastapi==0.116.1
uvicorn==0.35.0
psycopg[binary]==3.2.10
PyYAML==6.0.2
croniter==6.0.0
docker==7.1.0
jmespath==1.0.1
netaddr==1.3.0
proxmoxer==2.2.0
requests==2.32.5

